models/gemini_model.py
import google.generativeai as genai
from PIL import Image
from typing import Dict, Any, Optional
from .base_model import BaseVisionModel
from ..config import Config
import logging

logger = logging.getLogger(__name__)


class GeminiVisionModel(BaseVisionModel):
    """Google Gemini Vision model implementation"""

    def __init__(self):
        super().__init__("gemini-2.0-flash-exp")
        self.model = None

    def load_model(self) -> None:
        """Initialize Gemini model"""
        try:
            if not Config.GEMINI_API_KEY:
                raise ValueError("Gemini API key not found in environment variables")

            genai.configure(api_key=Config.GEMINI_API_KEY)

            # Configure model
            generation_config = {
                "temperature": 0.7,
                "top_p": 0.95,
                "top_k": 64,
                "max_output_tokens": 300,
            }

            self.model = genai.GenerativeModel(
                model_name="gemini-2.0-flash-exp",
                generation_config=generation_config,
            )

            self.is_loaded = True
            logger.info("Gemini model initialized successfully")

        except Exception as e:
            logger.error(f"Error initializing Gemini model: {str(e)}")
            raise

    def generate_description(
        self, image: Image.Image, prompt: Optional[str] = None
    ) -> str:
        """Generate description for the given image using Gemini"""
        if not self.is_loaded:
            self.load_model()

        try:
            # Default prompt for scene description
            if prompt is None:
                prompt = (
                    "Analyze this image and provide a detailed description. Include information about "
                    "people present, their activities, objects in the scene, the setting, and any "
                    "notable details or interactions you observe."
                )

            # Generate content
            response = self.model.generate_content([prompt, image])

            return response.text.strip()

        except Exception as e:
            logger.error(f"Error generating description with Gemini: {str(e)}")
            return f"Error: Unable to generate description - {str(e)}"

    def fine_tune(self, dataset_path: str, **kwargs) -> Dict[str, Any]:
        """Fine-tune Gemini model (placeholder - uses tuning API)"""
        try:
            # Note: Gemini fine-tuning requires specific dataset format and API calls
            logger.info("Gemini fine-tuning requires prepared dataset and API calls")

            return {
                "status": "info",
                "message": (
                    "Gemini fine-tuning requires dataset preparation and using "
                    "Google AI Studio or Vertex AI for model tuning. Please refer to "
                    "Google's documentation for detailed fine-tuning process."
                ),
            }

        except Exception as e:
            logger.error(f"Error with Gemini fine-tuning: {str(e)}")
            return {"status": "error", "message": f"Fine-tuning setup failed: {str(e)}"}
