requirements.txt
accelerate==0.24.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
attrs==25.3.0
bitsandbytes==0.41.3
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
datasets==2.14.6
dill==0.3.7
distro==1.9.0
dlib-binary==19.24.1
exceptiongroup==1.3.0
fastapi==0.104.1
filelock==3.18.0
frozenlist==1.6.0
fsspec==2023.10.0
google-ai-generativelanguage==0.4.0
google-api-core==2.25.0rc1
google-auth==2.40.2
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httpx==0.26.0
huggingface-hub==0.32.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.15
networkx==3.4.2
numpy==1.24.4
openai==1.3.5
opencv-python==4.8.1.78
packaging==25.0
pandas==2.2.3
peft==0.6.2
Pillow==10.1.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.7
psutil==7.0.0
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.5.0
pydantic_core==2.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.11.4
six==1.17.0
sniffio==1.3.1
starlette==0.27.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0+cu118
torchvision==0.22.0
tqdm==4.67.1
transformers==4.52.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.24.0
websockets==12.0
xxhash==3.5.0
yarl==1.20.0
